Add specific required feature for AR camera as required for Google Play Store submissions
Submitting an app to the Google Play Store based on this template fails with the following message:
"The uses-feature android.hardware.camera.ar is required in in the AndroidManifest.xml file when a dependency on ARCore is specified as required."

This commit adds the required line to the manifest to fix the error. An ARCore-based Amazon Sumerian app that I submitted to the Play Store passed after this change and went live.